requirements.txt
Django==2.0
djangorestframework==3.8.0

